Fix directory symlink creation.
Adds an optional "Type" parameter to New-SymLink. By default it is
"File" and causes the created symlink to be a file symlink. If
"Directory" is provided as an argument, the symlink is created with the
/D flag, indicating it is a directory symlink. If another value is
provided, an error is thrown thanks to [ValidateSet("...")].

Test-Path sometimes fails for files recently created by the git clone,
as if they didn't exist. If the spf13-vim/.vim directory failed the
check, the $HOME/.vimrc symlink was created without the /D flag,
breaking the installation.